Ugly dumpling
Disorganized dump of .R and .Rmd files pertaining to 2017-10 survey of DFO salmon staff.  
One set of files is for analysis of the survey directly.   
Another set is to extract tables from the survey result, and then convert complicated tables into nodes and links for neo4j. A lot of the work was ensuring compliance with RoundTable protocols for graph data.  
Some results in .html included; these have embedded results including graphics.
Some diagrams in .jpg and .pdf included.